\name{PredictModel-package}
\alias{PredictModel-package}
\title{Predict Model analysis toolkits}
\description{
Predict Model analysis toolkits
}
\usage{
PredictModelRs to Predict Model analysis toolkits
}
\arguments{
  \item{name}{a word}
}
\details{
  nothing
}
\value{
  no return
}
\references{
  nothing
}
