\name{PredictModelRs}
\alias{PredictModelRs}
\title{Predict for liver injury after admission and clinical status level for COVID-19 patients}
\usage{
  PredictModelRs(args)
}
\arguments{
\item{data}{a expression profile, test indexes in columns, samples in rows.}
\item{outrs}{TRUE or FLASE: TRUE output the results of }
}
\value{a dataframe, samples in rows, columns : output of the predictor(s)}

\description{
  A prediction model for COVID-19 patients
}
\author{liangyuan}
\keyword{internal}
\note{This is a contribution}
